coffee-bot/requirements.txt
requests
webexteamssdk
pillow